00-system/skills/integrations/google-gemini-image/pyproject.toml
[project]
name = "google-gemini-image"
version = "0.1.0"
description = "Gemini image generation skills for Nexus"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "Pillow>=10.0.0",
]

[tool.uv]
dev-dependencies = []
